Replace json to badger
