+++ Makefile
@@ -11,3 +11,8 @@ test/all:
 test/all/race:
 	$(GOTEST) -race $(TESTPKGS)
 
+update:
+	cd jpostcode-data && git pull origin master
+	go run script/badgerdump/main.go
+	go generate .
+
