+++ go.mod
@@ -3,6 +3,7 @@ module github.com/syumai/go-jpostcode
 go 1.13
 
 require (
+	github.com/dgraph-io/badger/v2 v2.2007.2 // indirect
 	github.com/goccy/go-json v0.1.15
 	github.com/google/go-cmp v0.5.4
 	github.com/mitchellh/mapstructure v1.1.2
